carga-service

+++ src/main/java/com/federicofrankenberger/tienda_api/repository/CategoriaRepository.java
@@ -3,5 +3,7 @@ package com.federicofrankenberger.tienda_api.repository;
 import com.federicofrankenberger.tienda_api.model.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categoria,Integer> {
+import java.util.Optional;
+
+public interface CategoriaRepository extends JpaRepository<Categoria,Long> {
 }
