+++ src/main/java/com/federicofrankenberger/tienda_api/repository/VentaRepository.java
@@ -3,5 +3,5 @@ package com.federicofrankenberger.tienda_api.repository;
 import com.federicofrankenberger.tienda_api.model.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VentaRepository extends JpaRepository<Venta,Integer> {
+public interface VentaRepository extends JpaRepository<Venta,Long> {
 }
